Fixed CMake with making libraries of SST
